Refactor dataset
